[Feature] USB Mass Storage Device
- This allows a USB Mass Storage device to expose the SD card.  Useful for taking off track logs.

NOTE:  I could not get the CDC mode to work once mass storage began, meaning, when you plug into the computer, it won't expose Serial, or ability to flash it.

To get around this limitation, I added a new platform (available in PlatformIO down the bottom bar) to enable a build to disable mass storage in non-release builds.