[FANET] Address Allocator
We have been given the 0x0C FANET vendor ID for Leaf.  To avoid collisions, we should start using them.

This adds a script to generate the NMV partition with an auto incremented  FANET address and store it in a csv file in the repo.

This should only be done before shipping them out by one person at the moment.